Litecoin: Reduce Dust thresholds by 10x
DUST_SOFT_LIMIT of 0.001 means a penalty of an additional mintxfee
is charged for each output smaller than 0.001.  This is a key
behavioral disincentive in Litecoin's anti-spam design.

DUST_HARD_LIMIT of 0.00001 means inputs smaller than this size are
ignored by the wallet, not available to coin selection and
effectively do not exist.  This too discourages spam.

Conflicts:
	src/main.h