Partial old pchMessageStart support
This is the end of the move from the old pchMessageStart to the new one started in the 0.13 release. This commit supports incoming connections from 0.13 clients sending dual pchMessageStart. 0.16 cannot talk directly to 0.9 clients as it does not fully support the old pchMessageStart. Once enough clients are using 0.16 we can consider using the new pchMessageStart in the next release with no support for the dual implementation in 0.13.